Use MPS if available
If the user doesn't specify the use of GPU's and the MPS device is
there, then use that device. It seems unlikely that a user on an M1/2
Mac that they will *not* want the MPS device used, but we can always
add a parameter later.

Rename `get_gpus_configuration` to `get_devices_configuration` since
the config is for CPU/GPU/MPS, not just GPUs.