banish remaining typing errors
We now only check the `run_mode` param once to determine whether we
create prediction or training sequences. The training sequence class
inherits from the prediction sequence class, and Sequence_T is a generic
type that allows us to accept lists of either one as parameters.

It's a bit unsatisfactory that the linter doesn't complain, since we're
now assuming that we have either train or prediction sequences when we
choose which collation function to use for constructing batches. I think
this has to do with the DataLoader class generic parameter ending up
as `Unknown` for some reason. I generally have trouble typing PyTorch.