Make sure dataclasses can be JSON-serialized
The advantage of using TypedDict was that it could be JSON-serialized
by default in Python. For dataclasses, we have to enable it explicitly
in the json.dumps() call. I wish Python had a proper mixin for this :/

Fixes this error:

TypeError: Object of type ModelParams_T is not JSON serializable