Encapsulate output indexing logic
Move all of the indexing and clone calls into a single method of
the batch output class; rename the batch output class so that the
"batch" aspect of it is clear, and have the function return a sentence
output class, whose typing structure so far looks exactly the same, but
the tensors contained are a different dimension. Man I wish we had
dependent types for working with Tensors!