Improve documentation and naming
* Converter->converter everywhere
* Document the data classes that are the core of our model.
* Note about how we are not doing the standard padding and masking and
why. I did a lot of experiments because I was sure that following the
documentation would lead to better performance, but it didn't, and I
don't really know why. I guess maybe because we're only annotating
individual words? But then, sometimes this depends on the overall
meaning of the sentence, so the CLS should be helpful in that case.