some comments and renames
Working on trying to understand all of the output dimensions. Add
comments giving the sizes for inputs to the scoring functions, as well
as explaining what the functions do. Some of this is still to do. Rename
some functions to be more clear about what they do, and use DUMMY_ID
instead of `<0` and `!=int(heads_true[0][0])` comparisons.