Use bool for start_subwords
Explain how start_subwords is used when constructing the input to the
Chu-Liu/Edmonds algorithm. Update comments everywhere, and use a bool
tensor instead of an int tensor to cut down on a few operations as well
as make its function clearer.

Properly generify the _pad_list function, as I had a bad bug where I
passed a bool list and an int pad, which resulted in an int Tensor
with true (DUMMY_ID) padding, leading to huge inputs to C-L/E and very
slow performance!