fix pyright errors
We have to ignore the typing errors for now because we can't properly
specify that `tree` is 1D in Python's type system, yet. I tried the
exploratory code at
https://taoa.io/posts/Shape-typing-numpy-with-pyright-and-variadic-generics
but couldn't get it to compile; pyright says
"TypeVarTuple not allowed in this context" to the `*Shape` part in
`NDArray = np.ndarray[*Shape, np.dtype[ScalarType]]`.