Reorganize model construction logic
* Previously fine-tuned models would either use the old annotation
schema or a new one generated from the new training data. Now they will
use a combined one by default.
* Add static constructors to BertForDepRel that handle model
construction under various circumstances.
* Use the resulting model configuration in UDDataset; this guarantees
that max_position_embeddings is set correctly.
* Pass the model into Predictor and Trainer as well. Now our constructor
parameters enforce that everything is constructed in order, and we don't
have to worry about strange errors occurring due to calling methods
in the wrong order.
* Remove conf parameter, replaced by model_path, pretrained_model_path,
and new_model_path. This is a breaking change :( It ensures that model/
config.json is always used for the model config, I believing simplifying
the config process. Parts of the config, such as the annotation schema,
should be seen as part of the model. Other parts are running params,
which I may extract later.