Encapsulate device in model class
Save the model's current device in a field, and send data to the device
before doing a forward pass. Now clients no longer need to specify the
device again when training, predicting or eval'ing.

It turned out that the special casing for `to()` during eval was
unnecessary, so remove it. Altogether this removes quite a bit of
complexity!

Support `to(device)`, as well.