train api and tests cleanup
Break up the e2e tests into smaller pieces, using pytest-order to make
sure that models are created before they are used.

Pass in plain sentences to the train() method and let the trainer call
the encoding method. Save the information about the dataset size in a
new data_description field in EvalResult.